// src/schemas/recipe-schema.ts
import { z } from 'zod';

export const CategorySchema = z.object({
    strCategory: z.string()
})

export const CategoryResponseSchema = z.object({
    drinks: z.array(CategorySchema)
})

export const SearchFilterSchema = z.object({
    ingredient: z.string(),
    category: z.string()
})


export const DrinkAPIResponseSchema = z.object({
    strDrink: z.string(),
    strDrinkThumb: z.string(),
    idDrink: z.string()
});

export const DrinksAPIResponseSchema = z.object({
    drinks: z.array(DrinkAPIResponseSchema)
});

export const RecipeAPIResponseSchema = z.object({
    idDrink: z.string(),
    strDrink: z.string(),
    strDrinkThumb: z.string(),
    strInstructions: z.string(),
    strIngredient1: z.string().nullable(),
    strIngredient2: z.string().nullable(),
    strIngredient3: z.string().nullable(),
    strIngredient4: z.string().nullable(),
    strIngredient5: z.string().nullable(),
    strIngredient6: z.string().nullable(),
    strMeasure1: z.string().nullable(),
    strMeasure2: z.string().nullable(),
    strMeasure3: z.string().nullable(),
    strMeasure4: z.string().nullable(),
    strMeasure5: z.string().nullable(),
    strMeasure6: z.string().nullable(),
});